+++ src/Nav.js
@@ -12,10 +12,10 @@ import AllPolls from "./Polls/AllPolls";
 import SinglePoll from "./Polls/SinglePoll";
 import CreatePoll from "./Polls/CreatePoll";
 import ForgotPassword from "./ForgotPassword";
-import Events from "./Events";
-import SingleEvent from "./SingleEvent";
-import AddEvent from "./AddEvent";
-import SingleEventAssigned from "./SingleEventAssigned";
+import Events from "./Events/Events";
+import SingleEvent from "./Events/SingleEvent";
+import AddEvent from "./Events/AddEvent";
+import SingleEventAssigned from "./Events/SingleEventAssigned";
 import AvatarGenerator from "./AvatarGenerator";
 import TwoUp from "./TwoUp";
